\name{IlluminaHumanMethylation450kanno.ilmn12.hg19}
\alias{IlluminaHumanMethylation450kanno.ilmn12.hg19}
\alias{Locations}
\alias{Other}
\alias{SNPs.Illumina}
\alias{Manifest}
\alias{Islands.UCSC}
\docType{data}
\title{
  Annotation data for the \sQuote{IlluminaHumanMethylation450k} micro array.
}
\description{
  The annotation data is based on Illumina's v1.2 annotation,
  specifically the file \sQuote{HumanMethylation450_15017482_v.1.2.csv}
  derived from \sQuote{HumanMethylation450_15017482_v.1.2.bpm}.
  Additional SNP annotation is generated by the authors.
}
\format{
  An object of class \code{IlluminaMethylationAnnotation} for
  \code{IlluminaHumanMethylation450kanno.ilmn12.hg19}. All others are of
  class \code{DataFrame}
}
\details{
  The following changes/ modifications / addition has been made to the
  source material.

  For the annotation related to UCSC islands, a value of \code{""} has
  been changed to \code{"OpenSea"}.

  The creation of this object based on the Illumina annotation and
  additional SNP information (see below) is contained in the
  \code{createAnno.R} script in the \code{scripts} directory.

  Contents of the columns are the responsibility of Illumina; see their
  documentation for details.
}

\source{
  See description.
}

\examples{
data(IlluminaHumanMethylation450kanno.ilmn12.hg19)
data(Locations)
data(Other)
data(Manifest)
data(SNPs.Illumina)
data(Islands.UCSC)
}
\keyword{datasets}
